refactor(topology): use marker interface for request and response types
